Combine all UpfDeploy types into one file
